LayoutTests/storage/indexeddb/modern/aborted-put-expected.txt

ALERT: First upgrade needed: Old version - 0 New version - 1
ALERT: [object IDBTransaction] - versionchange
ALERT: [object IDBDatabase]
ALERT: put succeeded - key was 'foo'
ALERT: get 'foo' succeeded - value was 'bar'
ALERT: get 'far' succeeded - value was 'undefined'
ALERT: First version change transaction completed
ALERT: Second upgrade needed: Old version - 1 New version - 2
ALERT: [object IDBTransaction] - versionchange
ALERT: [object IDBDatabase]
ALERT: get1 'foo' succeeded - value was 'baz'
ALERT: get2 'far' succeeded - value was 'boo'
ALERT: Second version change transaction abort
ALERT: Third upgrade needed: Old version - 1 New version - 2
ALERT: [object IDBTransaction] - versionchange
ALERT: [object IDBDatabase]
ALERT: get1 'foo' succeeded - value was 'bar'
ALERT: get2 'far' succeeded - value was 'undefined'
ALERT: Third version change transaction complete
ALERT: Done
This test creates some object stores, putting some values in them, committing the transaction.
In a new transaction, it then overwrites those values, but then aborts the transaction.
Finally it verifies everything is set up from the first transaction, and nothing from the aborted one committed.
